added react router and organized into pages
