+++ src/pages/HomePage.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function HomePage() {
+  return (
+    <h1>HomePage</h1>
+  )
+}
