+++ src/pages/LoginPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function LoginPage() {
+  return (
+    <h1>LoginPage</h1>
+  );
+}
